Moved classes responsible for autosend to a separate package
