+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -66,9 +66,9 @@ import org.odk.collect.android.gdrive.GoogleAccountPicker;
 import org.odk.collect.android.gdrive.GoogleAccountsManager;
 import org.odk.collect.android.gdrive.GoogleApiProvider;
 import org.odk.collect.android.geo.MapFragmentFactoryImpl;
-import org.odk.collect.android.instancemanagement.AutoSendSettingChecker;
-import org.odk.collect.android.instancemanagement.InstanceAutoSendFetcher;
-import org.odk.collect.android.instancemanagement.InstanceAutoSender;
+import org.odk.collect.android.instancemanagement.autosend.AutoSendSettingChecker;
+import org.odk.collect.android.instancemanagement.autosend.InstanceAutoSendFetcher;
+import org.odk.collect.android.instancemanagement.autosend.InstanceAutoSender;
 import org.odk.collect.android.itemsets.FastExternalItemsetsRepository;
 import org.odk.collect.android.logic.PropertyManager;
 import org.odk.collect.android.metadata.InstallIDProvider;
