+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/autosend/InstanceAutoSender.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.instancemanagement
+package org.odk.collect.android.instancemanagement.autosend
 
 import android.content.Context
 import org.odk.collect.analytics.Analytics
@@ -6,6 +6,8 @@ import org.odk.collect.android.R
 import org.odk.collect.android.formmanagement.InstancesAppState
 import org.odk.collect.android.gdrive.GoogleAccountsManager
 import org.odk.collect.android.gdrive.GoogleApiProvider
+import org.odk.collect.android.instancemanagement.InstanceSubmitter
+import org.odk.collect.android.instancemanagement.SubmitException
 import org.odk.collect.android.notifications.Notifier
 import org.odk.collect.android.projects.ProjectDependencyProvider
 import org.odk.collect.android.upload.FormUploadException
