+++ collect_app/src/test/java/org/odk/collect/android/instancemanagement/autosend/AutoSendSettingCheckerTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.instancemanagement
+package org.odk.collect.android.instancemanagement.autosend
 
 import android.net.ConnectivityManager
 import android.net.NetworkInfo
